[Fix] CSV-export of colspan/rowspans incorrect
When exporting a table containing a colspan followed by a rowspan as
a .csv-file, the rowspan will affect the wrong columns in following
rows. This commit adds a `current_colspan_for_row` to the
`ForEachVisibleCell`-function in order to keep track of the "amount of
colspan" in the current row so that the rowspan will be positioned
correctly in later rows.

There may still be some bugs in the way this is handled.